Add a domainKey property to Game class
The domainKey property is needed so that the GameManager is able to remove a game when it sends a close event. The GameManager holds its references to games in a hash of domainKeys mapped to arrays of games on the corresponding domains. To identify a game we need to know its domain (key) and its id.